rest-api company

+++ app/Http/Controllers/PlaceController.php
@@ -2,8 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\PlaceStoreRequest;
+use App\Http\Requests\PlaceUpdateRequest;
 use Illuminate\Http\Request;
 use App\Models\Place;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Validation\ValidationException;
 
 class PlaceController extends Controller
 {
@@ -26,46 +30,79 @@ class PlaceController extends Controller
      */
     public function show($id)
     {
-        $place = Place::findOrFail($id);
-        return response()->json($place);
+        $place = Place::find($id);
+        if (!empty($place)) {
+            return response()->json($place);
+        } else {
+            return response()->json([
+                "message" => "Place Not Found"
+            ], 404);
+        }
     }
 
     /**
      * Store a new place
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param  \App\Http\Requests\PlaceStoreRequest  $request
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function store(Request $request)
+    public function store(PlaceStoreRequest $request)
     {
-        $place = Place::create($request->all());
-        return response()->json($place, 201);
+        try {
+            $place = Place::create($request->validated());
+            return response()->json(["message" => "Place Added."], 201);
+        } catch (ValidationException $e) {
+            return response()->json([
+                'message' => 'Validation errors',
+                'errors' => $e->errors(),
+            ], 422);
+        }
     }
 
     /**
-     * Update a place
+     * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\Requests\PlaceUpdateRequest  $request
      * @param  string  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function update(Request $request, $id)
+    public function update(PlaceUpdateRequest $request, $id)
     {
-        $place = Place::findOrFail($id);
-        $place->update($request->all());
-        return response()->json($place);
+        try {
+            $place = Place::findOrFail($id);
+            $place->update($request->validated());
+            return response()->json(["message" => "Place updated successfully."], 200);
+        } catch (ValidationException $e) {
+            return response()->json([
+                'message' => 'Validation errors',
+                'errors' => $e->errors(),
+            ], 422);
+        } catch (\Exception $e) {
+            return response()->json(["message" => "Place not found."], 404);
+        }
     }
 
     /**
-     * Delete a place
+     * Delete a well
      *
-     * @param  string  $id
+     * @param string $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $place = Place::findOrFail($id);
-        $place->delete();
-        return response()->json(['message' => "Place with ID $id has been successfully deleted."], 200);
+        try {
+            $place = Place::findOrFail($id);
+
+            $relatedWellsCount = $place->wells()->where('placeId', $id)->count();
+
+            if ($relatedWellsCount > 0) {
+                return response()->json(['message' => 'Cannot delete this place because there are wells related to another place.'], 409);
+            }
+
+            $place->delete();
+            return response()->json(['message' => "Place with ID $id has been successfully deleted."], 200);
+        } catch (ModelNotFoundException $e) {
+            return response()->json(['message' => 'Place not found.'], 404);
+        }
     }
 }
